Refactored to Agenda and toString for Agenda
