+++ src/main/java/duke/logic/commands/DeleteCommand.java
@@ -4,7 +4,7 @@ import duke.logic.commands.results.CommandResultText;
 import duke.commons.exceptions.DukeException;
 import duke.commons.Messages;
 import duke.model.Model;
-import duke.model.events.Task;
+import duke.model.Task;
 
 /**
  * Class representing a command to delete a task.
