+++ src/main/java/duke/logic/commands/FindPathCommand.java
@@ -7,8 +7,8 @@ import duke.commons.Messages;
 import duke.commons.exceptions.DukeException;
 import duke.model.Model;
 import duke.model.lists.TaskList;
-import duke.model.events.Event;
-import duke.model.events.Task;
+import duke.model.Event;
+import duke.model.Task;
 import duke.model.locations.Venue;
 
 import java.util.ArrayList;
