+++ src/main/java/duke/logic/parsers/ParserStorageUtil.java
@@ -3,9 +3,8 @@ package duke.logic.parsers;
 import duke.commons.exceptions.DukeDateTimeParseException;
 import duke.commons.exceptions.DukeException;
 import duke.commons.Messages;
-import duke.model.events.Event;
-import duke.model.events.Task;
-import duke.model.planning.Agenda;
+import duke.model.Event;
+import duke.model.Task;
 import duke.model.planning.Todo;
 import duke.model.locations.BusStop;
 import duke.model.locations.TrainStation;
