+++ src/main/java/duke/ui/calendar/CalendarWindow.java
@@ -2,7 +2,7 @@ package duke.ui.calendar;
 
 import duke.logic.commands.results.CommandResultCalender;
 import duke.model.lists.EventList;
-import duke.model.events.Event;
+import duke.model.Event;
 import duke.ui.UiPart;
 import javafx.fxml.FXML;
 import javafx.scene.layout.GridPane;
