+++ src/test/java/duke/ModelStub.java
@@ -6,8 +6,8 @@ import duke.model.Model;
 import duke.model.lists.EventList;
 import duke.model.lists.TaskList;
 import duke.model.lists.VenueList;
-import duke.model.events.Event;
-import duke.model.events.Task;
+import duke.model.Event;
+import duke.model.Task;
 import duke.model.locations.BusStop;
 import duke.model.locations.Venue;
 import duke.model.planning.Agenda;
