+++ src/test/java/duke/logic/commands/MarkDoneCommandTest.java
@@ -2,7 +2,7 @@ package duke.logic.commands;
 
 import duke.ModelStub;
 import duke.commons.exceptions.DukeException;
-import duke.model.events.Task;
+import duke.model.Task;
 import duke.model.planning.Todo;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
